+++ actstream/tests/test_feeds.py
@@ -30,18 +30,20 @@ class FeedsTestCase(base.DataTestCase):
             'Two joined CoolGroup %s ago' % self.timesince,
         ]
         expected_json = expected[2:]
-        expected_json_with_user_activity = expected_json + ['admin started following Two %s ago' % self.timesince,
-        'admin joined CoolGroup %s ago' % self.timesince,
-        'admin commented on CoolGroup %s ago' % self.timesince
+        expected_times = [
+            'admin started following Two %s ago' % self.timesince,
+            'admin joined CoolGroup %s ago' % self.timesince,
+            'admin commented on CoolGroup %s ago' % self.timesince
         ]
+        expected_json_with_user_activity = expected_json + expected_times
         rss = self.capture('actstream_feed')
         self.assertAllIn(self.rss_base + expected, rss)
         atom = self.capture('actstream_feed_atom')
         self.assertAllIn(self.atom_base + expected, atom)
         json = self.capture('actstream_feed_json')
         self.assertAllIn(expected_json, json)
-        json_with_user_activity = self.capture('actstream_feed_json', query_string='with_user_activity=true') 
-        self.assertAllIn(expected_json_with_user_activity, json_with_user_activity) 
+        json_with_user_activity = self.capture('actstream_feed_json', query_string='with_user_activity=true')
+        self.assertAllIn(expected_json_with_user_activity, json_with_user_activity)
 
     def test_model_feed(self):
         expected = [
@@ -58,6 +60,7 @@ class FeedsTestCase(base.DataTestCase):
         atom = self.capture('actstream_model_feed_atom', self.user_ct.pk)
         self.assertAllIn(self.atom_base + expected, atom)
         json = self.capture('actstream_model_feed_json', self.user_ct.pk)
+        self.assertEqual(len(json['items']), 10)
 
     def test_object_feed(self):
         expected = [
@@ -74,3 +77,4 @@ class FeedsTestCase(base.DataTestCase):
             'actstream_object_feed_json',
             self.user_ct.pk, self.user2.pk
         )
+        self.assertEqual(len(json['items']), 3)